+++ services/announcements_service/src/oci_cli_announcements_service/announcement_cli_extended.py
@@ -44,11 +44,11 @@ announcement_cli.announce_root_group.add_command(announcements_group)
 announcement_cli.announce_root_group.add_command(user_status_group)
 announcements_group.add_command(announcement_cli.get_announcement)
 user_status_group.add_command(announcement_cli.update_announcement_user_status)
-user_status_group.add_command(announcement_cli.get_announcement_user_status)
+cli_util.rename_command(user_status_group, announcement_cli.get_announcement_user_status, "get")
 
 
 @cli_util.copy_params_from_generated_command(announcement_cli.list_announcements, params_to_exclude=['all_pages', 'limit', 'page_size'])
-@announcements_group.command(name=cli_util.override('list_announcements.command_name', 'list-announcements'), help="""Gets a list of `Announcement` objects for the current tenancy""")
+@announcements_group.command(name='list', help="""Gets a list of `Announcement` objects for the current tenancy""")
 @click.pass_context
 @json_skeleton_utils.json_skeleton_generation_handler(input_params_to_complex_types={}, output_type={'module': 'announcements_service', 'class': 'AnnouncementsCollection'})
 @cli_util.wrap_exceptions