+++ services/core/src/oci_cli_blockstorage/blockstorage_cli_extended.py
@@ -11,10 +11,12 @@ from oci_cli.cli_root import cli
 from oci_cli import cli_util
 from oci_cli import json_skeleton_utils
 
-
-cli.add_command(blockstorage_cli.blockstorage_root_group)
+cli_util.rename_command(cli, blockstorage_cli.blockstorage_root_group, "bv")
 blockstorage_cli.volume_group.commands.pop(blockstorage_cli.create_volume.name)
 blockstorage_cli.boot_volume_group.commands.pop(blockstorage_cli.create_boot_volume.name)
+cli_util.rename_command(blockstorage_cli.blockstorage_root_group, blockstorage_cli.volume_backup_group, "backup")
+blockstorage_cli.blockstorage_root_group.help = "Block Volume Service CLI"
+blockstorage_cli.blockstorage_root_group.short_help = "Block Volume Service"
 
 cli_util.update_param_help(blockstorage_cli.create_volume, 'availability_domain', """The Availability Domain of the volume. Example: `Uocm:PHX-AD-1`.