+++ tests/generated_test_extensions/extend_test_compute_management.py
@@ -1,0 +1,8 @@
+# coding: utf-8
+# Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+
+# These mappings are used by generated tests to look up operations that have been moved in code in the CLI.
+MOVED_COMMANDS = {
+    ('compute-management', 'instance', 'launch-compute-instance'): ['compute-management', 'instance-configuration', 'launch-compute-instance'],
+    ('compute-management', 'instance', 'list-instances'): ['compute-management', 'instance-pool', 'list-instances']
+}