+++ tests/generated_test_extensions/extend_test_iam.py
@@ -1,0 +1,7 @@
+# coding: utf-8
+# Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+
+# These mappings are used by generated tests to look up operations that have been moved in code in the CLI.
+MOVED_COMMANDS = {
+    ('iam', 'user_group_membership', 'add'): ['iam', 'group', 'add-user']
+}